+++ src/project/contract/eravm_assembly.rs
@@ -11,8 +11,6 @@ use crate::warning_type::WarningType;
 ///
 #[derive(Debug, serde::Serialize, serde::Deserialize, Clone)]
 pub struct Contract {
-    /// The EraVM version.
-    pub version: semver::Version,
     /// The contract source code.
     pub source_code: String,
 }
@@ -21,11 +19,8 @@ impl Contract {
     ///
     /// A shortcut constructor.
     ///
-    pub fn new(version: semver::Version, source_code: String) -> Self {
-        Self {
-            version,
-            source_code,
-        }
+    pub fn new(source_code: String) -> Self {
+        Self { source_code }
     }
 
     ///
@@ -35,6 +30,7 @@ impl Contract {
         self,
         contract_path: &str,
         metadata_hash: Option<era_compiler_common::Hash>,
+        no_bytecode_metadata: bool,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         llvm_options: Vec<String>,
         _output_selection: Vec<VyperSelector>,
@@ -54,9 +50,21 @@ impl Contract {
             debug_config.as_ref(),
         )?;
 
+        let cbor_data = if no_bytecode_metadata {
+            None
+        } else {
+            let cbor_key = crate::r#const::VYPER_PRODUCTION_NAME.to_owned();
+            let cbor_data = vec![(
+                crate::r#const::DEFAULT_EXECUTABLE_NAME.to_owned(),
+                crate::r#const::version().parse().expect("Always valid"),
+            )];
+            Some((cbor_key, cbor_data))
+        };
+
         let build = era_compiler_llvm_context::eravm_build(
             bytecode_buffer,
             metadata_hash,
+            cbor_data,
             Some(self.source_code),
         )?;