+++ src/project/contract/llvm_ir.rs
@@ -35,6 +35,7 @@ impl Contract {
         mut self,
         contract_path: &str,
         metadata_hash: Option<era_compiler_common::Hash>,
+        no_bytecode_metadata: bool,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         llvm_options: Vec<String>,
         output_selection: Vec<VyperSelector>,
@@ -62,9 +63,21 @@ impl Contract {
                 debug_config,
             );
 
+        let cbor_data = if no_bytecode_metadata {
+            None
+        } else {
+            let cbor_key = crate::r#const::VYPER_PRODUCTION_NAME.to_owned();
+            let cbor_data = vec![(
+                crate::r#const::DEFAULT_EXECUTABLE_NAME.to_owned(),
+                crate::r#const::version().parse().expect("Always valid"),
+            )];
+            Some((cbor_key, cbor_data))
+        };
+
         let build = context.build(
             contract_path,
             metadata_hash,
+            cbor_data,
             output_selection.contains(&VyperSelector::EraVMAssembly),
             false,
         )?;