+++ src/project/contract/vyper/expression/mod.rs
@@ -94,7 +94,8 @@ impl Expression {
     ///
     pub fn safe_label(label: &str) -> String {
         let identifier = label.replace(['(', ')', '[', ']', ',', ' '], "_");
-        let hash = era_compiler_common::Hash::keccak256(identifier.as_bytes()).to_string();
+        let hash =
+            era_compiler_common::Keccak256Hash::from_slice(identifier.as_bytes()).to_string();
         format!(
             "{identifier}_{}",
             hash.strip_prefix("0x").unwrap_or(hash.as_str())